+++ primaryHandler.go
@@ -2,26 +2,37 @@ package main
 
 import (
 	"bytes"
+	"context"
 	"encoding/base64"
 	"errors"
 	"fmt"
 	"net/http"
+	"os"
+	"os/signal"
 	"regexp"
+	"syscall"
 
+	"emperror.dev/emperror"
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	"github.com/gorilla/mux"
 	"github.com/justinas/alice"
+	"github.com/prometheus/client_golang/prometheus"
 	"github.com/spf13/viper"
 	"github.com/xmidt-org/ancla"
 	"github.com/xmidt-org/bascule"
 	bchecks "github.com/xmidt-org/bascule/basculechecks"
 	"github.com/xmidt-org/bascule/basculehttp"
-	"github.com/xmidt-org/bascule/key"
+	"github.com/xmidt-org/clortho"
+	"github.com/xmidt-org/clortho/clorthometrics"
+	"github.com/xmidt-org/clortho/clorthozap"
+	"github.com/xmidt-org/sallust"
+	"github.com/xmidt-org/touchstone"
 	"github.com/xmidt-org/webpa-common/v2/basculechecks"
 	"github.com/xmidt-org/webpa-common/v2/basculemetrics"
 	"github.com/xmidt-org/webpa-common/v2/logging"
 	"github.com/xmidt-org/webpa-common/v2/xmetrics"
+	"go.uber.org/zap"
 )
 
 const (
@@ -40,8 +51,8 @@ type CapabilityConfig struct {
 
 // JWTValidator provides a convenient way to define jwt validator through config files
 type JWTValidator struct {
-	// Keys is used to create the key.Resolver for JWT verification keys
-	Keys key.ResolverFactory `json:"key"`
+	// Config is used to create the clortho Resolver & Refresher for JWT verification keys
+	Config clortho.Config `json:"config"`
 
 	// Leeway is used to set the amount of time buffer should be given to JWT
 	// time values, such as nbf
@@ -102,21 +113,83 @@ func authenticationMiddleware(v *viper.Viper, logger log.Logger, registry xmetri
 	}
 
 	var jwtVal JWTValidator
+	// Get jwt configuration, including clortho's configuration
 	v.UnmarshalKey("jwtValidator", &jwtVal)
-	if jwtVal.Keys.URI != "" {
-		resolver, err := jwtVal.Keys.NewResolver()
-		if err != nil {
-			return &alice.Chain{}, fmt.Errorf("failed to create resolver: %v", err)
-		}
+	kr := clortho.NewKeyRing()
 
-		options = append(options, basculehttp.WithTokenFactory("Bearer", basculehttp.BearerTokenFactory{
-			DefaultKeyID: defaultKeyID,
-			Resolver:     resolver,
-			Parser:       bascule.DefaultJWTParser,
-			Leeway:       jwtVal.Leeway,
-		}))
+	// Instantiate a fetcher for refresher and resolver to share
+	f, err := clortho.NewFetcher()
+	if err != nil {
+		return &alice.Chain{}, emperror.With(err, "failed to create clortho fetcher")
 	}
 
+	ref, err := clortho.NewRefresher(
+		clortho.WithConfig(jwtVal.Config),
+		clortho.WithFetcher(f),
+	)
+	if err != nil {
+		return &alice.Chain{}, emperror.With(err, "failed to create clortho refresher")
+	}
+
+	resolver, err := clortho.NewResolver(
+		clortho.WithConfig(jwtVal.Config),
+		clortho.WithKeyRing(kr),
+		clortho.WithFetcher(f),
+	)
+	if err != nil {
+		return &alice.Chain{}, emperror.With(err, "failed to create clortho resolver")
+	}
+
+	promReg, ok := registry.(prometheus.Registerer)
+	if !ok {
+		return &alice.Chain{}, errors.New("failed to get prometheus registerer")
+	}
+
+	var (
+		tsConfig touchstone.Config
+		zConfig  sallust.Config
+	)
+	// Get touchstone & zap configurations
+	v.UnmarshalKey("touchstone", &tsConfig)
+	v.UnmarshalKey("zap", &zConfig)
+	zlogger := zap.Must(zConfig.Build())
+	tf := touchstone.NewFactory(tsConfig, zlogger, promReg)
+	// Instantiate a metric listener for refresher and resolver to share
+	cml, err := clorthometrics.NewListener(clorthometrics.WithFactory(tf))
+	if err != nil {
+		return &alice.Chain{}, emperror.With(err, "failed to create clortho metrics listener")
+	}
+
+	// Instantiate a logging listener for refresher and resolver to share
+	czl, err := clorthozap.NewListener(
+		clorthozap.WithLogger(zlogger),
+	)
+	if err != nil {
+		return &alice.Chain{}, emperror.With(err, "failed to create clortho zap logger listener")
+	}
+
+	resolver.AddListener(cml)
+	resolver.AddListener(czl)
+	ref.AddListener(cml)
+	ref.AddListener(czl)
+	ref.AddListener(kr)
+	// context.Background() is for the unused `context.Context` argument in refresher.Start
+	ref.Start(context.Background())
+	// Shutdown refresher's goroutines when SIGTERM
+	sigs := make(chan os.Signal, 1)
+	signal.Notify(sigs, syscall.SIGTERM)
+	go func() {
+		<-sigs
+		// context.Background() is for the unused `context.Context` argument in refresher.Stop
+		ref.Stop(context.Background())
+	}()
+
+	options = append(options, basculehttp.WithTokenFactory("Bearer", basculehttp.BearerTokenFactory{
+		DefaultKeyID: defaultKeyID,
+		Resolver:     resolver,
+		Parser:       bascule.DefaultJWTParser,
+		Leeway:       jwtVal.Leeway,
+	}))
 	authConstructor := basculehttp.NewConstructor(append([]basculehttp.COption{
 		basculehttp.WithParseURLFunc(basculehttp.CreateRemovePrefixURLFunc("/"+apiBase+"/", basculehttp.DefaultParseURLFunc)),
 	}, options...)...)
@@ -124,7 +197,6 @@ func authenticationMiddleware(v *viper.Viper, logger log.Logger, registry xmetri
 		basculehttp.WithParseURLFunc(basculehttp.CreateRemovePrefixURLFunc("/api/"+prevAPIVersion+"/", basculehttp.DefaultParseURLFunc)),
 		basculehttp.WithCErrorHTTPResponseFunc(basculehttp.LegacyOnErrorHTTPResponse),
 	}, options...)...)
-
 	bearerRules := bascule.Validators{
 		bchecks.NonEmptyPrincipal(),
 		bchecks.NonEmptyType(),